requirements.txt
sgp4
pandas
numpy
pyarrow
fastparquet
spacetrack
scikit-learn
torch
fastapi
uvicorn
python-dotenv
httpx
scipy
websockets